refactor: WinningLotto 생성, 당첨로직 구조 변경
