+++ src/main/java/lotto/domain/WinResult.java
@@ -3,18 +3,15 @@ package lotto.domain;
 import java.util.Map;
 
 public class WinResult {
-    private int purchasedMoney;
+    private final int purchasedMoney;
     private final Lottos purchasedLotto;
-    private final Lotto winNumbers;
-    private final LottoNumber bonusNumber;
+    private final WinningLotto winningLotto;
     private final Map<WinType, Integer> winResultMap;
-    private final WinningResultReader winningResultReader = new WinningResultReader();
 
-    public WinResult(int purchasedMoney, Lottos purchasedLotto, Lotto winNumbers, LottoNumber bonusNumber) {
+    public WinResult(int purchasedMoney, Lottos purchasedLotto, WinningLotto winningLotto) {
         this.purchasedMoney = purchasedMoney;
         this.purchasedLotto = purchasedLotto;
-        this.winNumbers = winNumbers;
-        this.bonusNumber = bonusNumber;
+        this.winningLotto = winningLotto;
         this.winResultMap = getWinResultMap();
     }
 
@@ -28,6 +25,6 @@ public class WinResult {
     }
 
     private Map<WinType, Integer> getWinResultMap() {
-        return winningResultReader.getWinningResult(purchasedLotto, winNumbers, bonusNumber);
+        return winningLotto.matchWinning(purchasedLotto);
     }
 }
